requirements.txt
# Core dependencies
pyttsx3~=2.90
wikipedia~=1.4.0
PyAutoGUI~=0.9.54
SpeechRecognition~=3.10.4
pyjokes==0.6.0

# AI Integration
openai>=1.0.0
google-generativeai>=0.3.0

# System Control & Monitoring
psutil>=5.9.0
pycaw>=20220416
screen-brightness-control>=0.20.0

# Communication & Automation
pywhatkit>=5.4
requests>=2.31.0

# Configuration & Scheduling
python-dotenv>=1.0.0
APScheduler>=3.10.0

# News
newsapi-python>=0.2.7

# Music & Entertainment
spotipy>=2.23.0
youtube-search-python>=1.6.6

# Weather (optional)
pyowm>=3.3.0